feat: Started working on my model classes
I've created a Sections class that will be the parent class
for each section of my resume. For now, I've also created a Skills
class which inherits from Section that has a companion object
of all of my skills. I now need to think about how I'm going to
display the skills within the RV that I've created.