feat: Created a nested RV for my Skill model
For the CardView that displays a Skill, I want to show a nested
RV. The first RV will be for each section of my resume. Then,
within each section, I want to show headers and bullets for
each header. For now, I've created the headers within the Skills
section. Next task: show the actual bullets for each header.