feat: Created Collapsible/Expandable CardViews
After spending a few days frantically trying to learn about
animations, I decided to throw in the towel and do something
much simpler. For now, I am simply hiding/showing the RV
that displays the contents within each CardView. I've added arrows
beside each title as a helpful indicator of whether the contents
are shown or not.

Later on, I'd love to add an animation to all of this so that it
is shown in a much smoother way.