feat: Now displaying all projects within the fragment
I initially thought it'd be a bad idea if I showed all of my
projects on the screen, thinking it'd fill up the screen too
much. Therefore, I was only showing one project, with a "More"
button to display all of the projects if the user so wished. I've
since decided against this idea and have decided I will just
display all of my projects with the balance of the expandable/
collapsible CardView extension functions that I've created.

Next step is to remove the AllProjectsFragment entirely since it
is no longer needed.