feat: Created UI for resume PDF
So far, I've added an Adobe Acrobat icon to the bottom of my
resume, which is intended to allow the user to view my resume
in PDF form within another app. There is some partial work
already in place for accessing the file, but that's all still
a work in progress.