feat: Resume PDF can now be opened by external apps
This was a bit more complex than I had originally anticipated.
First, I created an extension function to Resources so that I
can easily access the raw resource of the PDF. Then, I had to
create a FileProvider so that the file can be safely shared with
external apps. The rest was simple: define an on-click for the
Adobe icon. Within the on-click, define an Intent which uses
the PDF file as its data w/ appropriate action and MIME type
defined and start the external app.