feat: Work in progress with animations
I've added the functionality for an animation on my CardViews,
but things get really wacky. If you click on the first CardView,
then the animation works although there is some small delay.
However, on the remaining CardViews, the animation does not work
correctly. It appears to start, but then the CardView never
collapses (it's always expanded). For now, I wanted to keep all
of the work that I've done thus far so that I don't lose it.