feat: Animations are now working, hooray!
It turns out that the major issue that I was having with the
animations was due to the fact that I was using a ConstraintLayout
within the containing CardView. When the CardView was collapsed,
the RecyclerView (child of the ConstraintLayout) had a height of
0. This caused the RecyclerView to match constraints, since
that's what occurs with any View within a ConstraintLayout
with a height (or width) or 0.

To get around this, I added a listener to my animations for the
end and start which sets the visibility of the RecyclerView
to GONE or VISIBLE respectively. This has solved the animation
problems that I previously had. Awesome!